merge/pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
description = "HTTP Frontend as a django app for argus-server using HTMx"
name = "argus-htmx-frontend"
readme = "README.rst"
authors = [{name = "Hanne Moa", email = "hanne.moa@sikt.no"}]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Programming Language :: Python :: 3 :: Only",
    "Environment :: Web Environment",
]
dynamic = ["version"]
dependencies = [
    "argus-server>=1.27.0",
    "django-htmx",
    "django-widget-tweaks==1.5.0",
    "social-auth-core>=4.1",
    "social-auth-app-django>=5.0",
]

[project.urls]
Github = "https://github.com/Uninett/argus-htmx-frontend"

[project.optional-dependencies]
docs = ["sphinx"]

[tool.flit.module]
name = "argus_htmx"

[tool.ruff]
line-length = 120
output-format = "full"

[tool.djlint]
profile="django"
indent = 2
ignore="H006"
